Extract logVNode
