+++ src/adapter/10/renderer.ts
@@ -43,6 +43,7 @@ import {
 	createVNodeId,
 	updateVNodeId,
 } from "./IdMapper";
+import { logVNode } from "./renderer/logVNode";
 
 export interface RendererConfig10 {
 	Fragment: FunctionalComponent;
@@ -136,14 +137,7 @@ export function createRenderer(
 				if (c) c.forceUpdate();
 			}
 		},
-		log(id, children) {
-			const vnode = getVNodeById(ids, id);
-			if (vnode == null) {
-				console.warn(`Could not find vnode with id ${id}`);
-				return;
-			}
-			logVNode(vnode, id, children, config);
-		},
+		log: (id, children) => logVNode(ids, config, id, children),
 		inspect(id) {
 			const vnode = getVNodeById(ids, id);
 			if (!vnode) return null;
@@ -303,32 +297,6 @@ export function createRenderer(
 			}
 		},
 	};
-}
-
-/**
- * Print an element to console
- */
-export function logVNode(
-	vnode: VNode,
-	id: ID,
-	children: ID[],
-	config: RendererConfig10,
-) {
-	const display = getDisplayName(vnode, config);
-	const name = display === "#text" ? display : `<${display || "Component"} />`;
-
-	/* eslint-disable no-console */
-	console.group(`LOG %c${name}`, "color: #ea88fd; font-weight: normal");
-	console.log("props:", vnode.props);
-	const c = getComponent(vnode);
-	if (c != null) {
-		console.log("state:", c.state);
-	}
-	console.log("vnode:", vnode);
-	console.log("devtools id:", id);
-	console.log("devtools children:", children);
-	console.groupEnd();
-	/* eslint-enable no-console */
 }
 
 export function createCommit(
